backend/fastapi/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
threadpoolctl==3.5.0
tldextract==5.1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
